Refactor SEG-Y inference utilities into a separate module
